+++ changelog.md
@@ -8,6 +8,7 @@
 - Added models for some [My Nether's Delight](https://www.curseforge.com/minecraft/mc-mods/my-nethers-delight) items on [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin) Coasters
 - Added models for some [Abnormals Delight](https://www.curseforge.com/minecraft/mc-mods/abnormals-delight) items on [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin) Coasters
 - Added models for some [Undergarden Delight](https://www.curseforge.com/minecraft/mc-mods/undergarden-delight) items on [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin) Coasters
+- Glow Jam Cookie will now provide the `Tracer` effect in presence of [Cosmopolitan](https://www.curseforge.com/minecraft/mc-mods/cosmopolitan)
 - Changed ingredient in grass cutting recipe from `#straw_plants` to Grass
 - Added recipes for cutting several modded grasses into Straw
 - Added Clover from [Regions Unexplored](https://www.curseforge.com/minecraft/mc-mods/regions-unexplored) to `#forge:clover`, allowing it to be cut into Chopped Clover